refactor tuf pkg to ease development
