+++ pkg/tuf/common_test.go
