+++ scripts/signy-list.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
+source scripts/signy-env.sh
+
 # FIXME: list does not seem to work right now
-signy --tlscacert=$GOPATH/src/github.com/theupdateframework/notary/cmd/notary/root-ca.crt --server=https://localhost:4443 --log=info list localhost:5000/thin-bundle:v1
+run_signy list
