+++ scripts/signy-sign.sh
@@ -2,4 +2,6 @@
 
 source scripts/signy-env.sh
 
-signy --tlscacert=$GOPATH/src/github.com/theupdateframework/notary/cmd/notary/root-ca.crt --server=https://localhost:4443 --log=info sign testdata/cnab/bundle.json localhost:5000/thin-bundle:v1
+# $1: bundle.json
+# $2: GUN
+run_signy sign $1 $2
