+++ scripts/signy-verify.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-signy --tlscacert=$GOPATH/src/github.com/theupdateframework/notary/cmd/notary/root-ca.crt --server=https://localhost:4443 --log=info verify localhost:5000/thin-bundle:v1
+source scripts/signy-env.sh
+
+# $1: GUN
+run_signy verify $1
